ticket tag update and delete action added
